feat(cli): record report provenance through the kernel dispatch
The recorder held a host-side mapping for the report family, and it
restated kernel builders. Kernel 0.6.0 owns the family now. Thus every
prov member walks one path, and the exhaustiveness guard covers the
whole contract.

One bridge realizes the whole provenance seam of the harness. The block
members carry the block kind, and the kernel pin moves to 0.6.0 in the
same commit as the rework.
