+++ cli/openspec/changes/record-report-provenance/tasks.md
@@ -1,28 +1,25 @@
 # Tasks — record-report-provenance
 
-## 1. The types and the bus
+## 1. The kernel pin and the types
 
-- [x] 1.1 Add the report ref shapes to `src/types/prov.ts`, as cli-owned types
-- [x] 1.2 Add the nine report members to the `BusEvent` union in `src/types/events.ts`
-- [x] 1.3 Add the telemetry projection of each member to `eventFields` in `src/lib/bus.ts`
+- [x] 1.1 Move the `@inflexa-ai/prov-kernel` pin to `^0.6.0`
+- [x] 1.2 Replace the cli-owned report ref types in `src/types/prov.ts` with kernel re-exports
+- [x] 1.3 Add `blockKind` to the four block members in `src/types/events.ts`, and to their `eventFields` projection in `src/lib/bus.ts`
 
-## 2. The recorder mapping
+## 2. The uniform recorder
 
-- [x] 2.1 Branch the report family in `onEvent`, before `toKernelEvent`
-- [x] 2.2 Mint the report QName cli-side, over `cliProvDigest` of the thread id
-- [x] 2.3 Map the session creation: the report entity with the parent attribute, or the conversation action
-- [x] 2.4 Map each act as one typed lifecycle action with the thread id and its data
-- [x] 2.5 Map the version record: the version entity, attributed to the report entity
-- [x] 2.6 Guard the mapping: a throw logs and returns, and the emitter never sees it
-- [x] 2.7 Cover the mapping with targeted recorder tests, on the flushed document state
+- [x] 2.1 Extend `toKernelEvent` with the nine report members
+- [x] 2.2 Delete the host mapping from `src/modules/prov/prov.ts`: `isReportEvent`, `appendReportRecords`, and the helper set
+- [x] 2.3 Point the recorder fixture tests at the kernel arms, and keep the flushed-document assertions unchanged
+- [x] 2.4 Cover the double-emit guard on the flushed document state
 
-## 3. The two seam realizations
+## 3. The merged bridge
 
-- [x] 3.1 Make `src/modules/harness/report_bridge.ts`: the seam events onto the bus members, with the system actor and the live session model
-- [x] 3.2 Realize `readReportProvenance`: drain the flush, read the stored bytes, build the attestation
-- [x] 3.3 Bind the two seams on the core bag in `src/modules/harness/runtime.ts`
-- [x] 3.4 Pass the bound seam into `PrepareChatTurnDeps` at the chat-turn call site
-- [x] 3.5 Cover both realizations with targeted tests, absence included
+- [x] 3.1 Fold `report_bridge.ts` into `src/modules/harness/prov_bridge.ts`, with the session emit, the live-model source, and trimmed comments
+- [x] 3.2 Carry `blockKind` through the session emit mapping
+- [x] 3.3 Move the read realization into the merged bridge, unchanged in behavior
+- [x] 3.4 Bind one `ProvenanceSeam` object on the core bag in `src/modules/harness/runtime.ts`, and pass it into `PrepareChatTurnDeps`
+- [x] 3.5 Update the bridge tests to the merged module, absence included
 
 ## 4. Verification
