+++ cli/package.json
@@ -39,7 +39,7 @@
         "@clack/core": "1.4.3",
         "@clack/prompts": "^1.7.0",
         "@inflexa-ai/harness": "0.31.1",
-        "@inflexa-ai/prov-kernel": "^0.5.0",
+        "@inflexa-ai/prov-kernel": "^0.6.0",
         "@inflexa-ai/tsprov": "0.5.1",
         "@opentelemetry/api": "^1.9.1",
         "@opentelemetry/api-logs": "^0.219.0",
