+++ cli/src/modules/harness/prov_bridge.ts
@@ -1,20 +1,38 @@
-import type { ArtifactRegistrationInput, ArtifactRegistry, ExternalRegistrationResult, RunProvenanceEvent } from "@inflexa-ai/harness";
+import type {
+    ArtifactRegistrationInput,
+    ArtifactRegistry,
+    ExternalRegistrationResult,
+    ProvenanceExport,
+    ProvenanceSeam,
+    RunProvenanceEvent,
+    SessionProvenanceEvent,
+} from "@inflexa-ai/harness";
 
+import { getAnalysisProvenance } from "../../db/primary_query.ts";
 import { Bus } from "../../lib/bus.ts";
+import { getLogger } from "../../lib/log.ts";
 import type { ProvCommandInputRef, ProvCommandRef, ProvFileKey, ProvFileRef, ProvModelId, ProvStepRef, ProvUsedInputRef } from "../../types/prov.ts";
 import { provModel } from "../prov/document.ts";
-import { systemActor } from "../prov/prov.ts";
+import { flushProvenanceAsync, systemActor } from "../prov/prov.ts";
+import { buildAttestation } from "../prov/verify.ts";
 
-// The cli↔harness provenance bridge: the two halves that connect the harness's execution machinery
-// to the cli's signed provenance ledger, kept together because both translate harness execution
-// facts into `prov.*` bus events and nothing else. The recorder (`modules/prov/prov.ts`) subscribes
-// to those events and appends to the in-memory tsprov document.
+// The cli↔harness provenance bridge: the whole `ProvenanceSeam` over ONE signed analysis document.
+// The run emit and the artifact registry translate execution facts into `prov.*` bus events, the
+// session emit does the same for each act of a session, and the document read gives the recorded
+// bytes back for the page of a preview. The recorder (`modules/prov/prov.ts`) subscribes to those
+// events and appends to the in-memory tsprov document.
+//
+// The four halves share a file because they share a subject: the read hands out exactly what the
+// emits recorded, thus a change of the record family meets its reader in the same place. The run
+// bridge (`run_bridge.ts`) stays apart for the opposite reason — a lossy presentation channel and a
+// signed hash chain answer to different masters, and one file would invite one payload.
 //
 // This module is the ONLY place a run-engine artifact/input observation crosses into the cli's
-// provenance vocabulary. It imports nothing DB-related: it only emits bus events, so it satisfies
-// the `ArtifactRegistry` seam's "MUST NOT touch cortex_artifacts" contract by construction — the
-// harness owns that local-ledger write AROUND this seam, and writes the returned `externalId` back
-// onto its row itself.
+// provenance vocabulary. The registry reads no store at all: it only emits bus events, so it
+// satisfies the `ArtifactRegistry` seam's "MUST NOT touch cortex_artifacts" contract by construction
+// — the harness owns that local-ledger write AROUND this seam, and writes the returned `externalId`
+// back onto its row itself. The one store read in this file is the document read at the bottom,
+// which reaches the cli's OWN analyses table and never the harness ledger.
 //
 // The step-lifecycle split: this adapter emits COMMAND, FILE, and USED-INPUT events — the finer-grained
 // command lineage plus per-file generations and per-input reads. Step activities (`prov.step_completed`)
@@ -402,4 +420,201 @@ export function createSwappableSandboxEmitters(initialName: ProvModelId): Swappa
             emitter = createRunProvenanceEmitter(name);
         },
     };
+}
+
+const log = getLogger("harness:prov");
+
+/**
+ * The live provenance name of the model that drives a session, as `{provider}/{model}`.
+ *
+ * A function and not a value, because a person can change the model of the conversation agent while
+ * the runtime runs. The session emit reads the name at each act, thus every record names the model
+ * that did it.
+ */
+export type ReportSessionModel = () => ProvModelId;
+
+// The one installed source. A process runs one embedded runtime, thus one source. It is a module
+// singleton and not a parameter of the emit, because the composition root and the chat-turn engine
+// must reach ONE seam: a second binding would be a second, and possibly different, claim about the
+// model.
+let sessionModel: ReportSessionModel | null = null;
+
+/**
+ * Install the source of the live session model, or clear it with `null`.
+ *
+ * The boot installs it beside the agent switch that it reads. A second install replaces the first,
+ * thus a re-boot needs no teardown of its own.
+ */
+export function installReportSessionModel(read: ReportSessionModel | null): void {
+    sessionModel = read;
+}
+
+/** The four block acts name four different bus members, and they carry one payload shape. */
+type ReportBlockMember = "prov.report_block_added" | "prov.report_block_changed" | "prov.report_block_removed" | "prov.report_block_moved";
+
+/**
+ * Realize the session emit of the seam as bus emission: map each observed act of a session onto its
+ * `prov.*` member, stamped with the system actor and the live session model.
+ *
+ * The agent does each of these acts, thus the record names the system actor and carries the model on
+ * whose behalf it acted — the same treatment a model-driven step gets. The model is read at emit
+ * time, thus a live agent switch re-stamps every later act.
+ *
+ * The mapping restates each payload into the cli's own ref names instead of forwarding the harness
+ * object. Thus a widened seam shape fails to compile here, and it never reaches a subscriber
+ * unannounced.
+ */
+export function emitSessionObservation(event: SessionProvenanceEvent): void {
+    const live = sessionModel?.();
+    if (live === undefined) {
+        // A report record names the model that did the act. With no live name the bridge cannot make
+        // that claim, and an invented name in a signed document is worse than a missing record.
+        log.warn({ event: event.type, analysisId: event.analysisId }, "no live session model; report record dropped");
+        return;
+    }
+    // One name for the whole mapping, declared without the absent half, because the block helper
+    // below reads it inside its own body and a narrowing does not reach there.
+    const model: ProvModelId = live;
+    const actor = systemActor();
+    const analysisId = event.analysisId;
+
+    // The thread, the block, and its kind ride in as arguments, because the narrowing of the switch
+    // below does not reach into this body.
+    function emitBlockAct(type: ReportBlockMember, threadId: string, blockId: string, blockKind: string): void {
+        Bus.emit("inflexa", { type, analysisId, actor, model, block: { threadId, blockId, blockKind } });
+    }
+
+    switch (event.type) {
+        case "create-session":
+            Bus.emit("inflexa", {
+                type: "prov.session_created",
+                analysisId,
+                actor,
+                model,
+                session: {
+                    threadId: event.threadId,
+                    kind: event.sessionKind,
+                    // A root session has no parent, thus the key stays absent rather than carrying
+                    // `undefined` into the record.
+                    ...(event.parentThreadId !== undefined ? { parentThreadId: event.parentThreadId } : {}),
+                },
+            });
+            return;
+        case "add-block":
+            emitBlockAct("prov.report_block_added", event.threadId, event.blockId, event.blockKind);
+            return;
+        case "change-block":
+            emitBlockAct("prov.report_block_changed", event.threadId, event.blockId, event.blockKind);
+            return;
+        case "remove-block":
+            emitBlockAct("prov.report_block_removed", event.threadId, event.blockId, event.blockKind);
+            return;
+        case "move-block":
+            emitBlockAct("prov.report_block_moved", event.threadId, event.blockId, event.blockKind);
+            return;
+        case "set-title":
+            Bus.emit("inflexa", { type: "prov.report_title_set", analysisId, actor, model, title: { threadId: event.threadId, title: event.title } });
+            return;
+        case "run-derivation":
+            Bus.emit("inflexa", {
+                type: "prov.report_derivation_run",
+                analysisId,
+                actor,
+                model,
+                derivation: {
+                    threadId: event.threadId,
+                    outputPath: event.outputPath,
+                    outputHash: event.outputHash,
+                    scriptHash: event.scriptHash,
+                    // Copied pair by pair, so nothing that the tool still holds reaches a subscriber.
+                    sources: event.sources.map((s) => ({ path: s.path, hash: s.hash })),
+                },
+            });
+            return;
+        case "preview":
+            Bus.emit("inflexa", {
+                type: "prov.report_previewed",
+                analysisId,
+                actor,
+                model,
+                preview: { threadId: event.threadId, pagePath: event.pagePath, documentHash: event.documentHash },
+            });
+            return;
+        case "record-version":
+            Bus.emit("inflexa", {
+                type: "prov.report_version_recorded",
+                analysisId,
+                actor,
+                model,
+                version: { threadId: event.threadId, versionId: event.versionId, replaced: event.replaced },
+            });
+            return;
+        default: {
+            // Exhaustiveness: a new session event must add its mapping here.
+            const never: never = event;
+            throw new Error(`unhandled session provenance event: ${JSON.stringify(never)}`);
+        }
+    }
+}
+
+/**
+ * Realize the document read of the seam: give the stored provenance document of an analysis, with a
+ * fresh attestation over the exact bytes that the page carries.
+ *
+ * The recorder writes the column on a debounced flush, thus a preview that reads at once would miss
+ * the acts of the session that asks for it. The drain closes that window, and it costs one flush for
+ * each preview.
+ *
+ * The read goes to the stored column, and not through `serializeProvenance`. That helper seeds a
+ * fresh document when the column is unset, and it then gives bytes that no signature covers. A
+ * document with no proof must not reach a page.
+ *
+ * Absence is a normal result and never an error: a vanished analysis row, an analysis whose column
+ * is still null, and a failed attestation build all give nothing, and the page then renders with no
+ * provenance asset. The last of the three is a fault, thus it also reaches the log.
+ */
+export async function readProvenanceExport(analysisId: string): Promise<ProvenanceExport | undefined> {
+    await flushProvenanceAsync();
+
+    const document = getAnalysisProvenance(analysisId).match(
+        (json) => json,
+        (e) => {
+            log.error({ analysisId, err: e.type, cause: e.cause }, "failed to read the stored provenance; the page gets none");
+            return null;
+        },
+    );
+    if (document === null) return undefined;
+
+    const attestation = await buildAttestation(document);
+    if (attestation.isErr()) {
+        log.error({ analysisId, err: attestation.error.type }, "failed to build the attestation; the page gets no provenance");
+        return undefined;
+    }
+    return { document, attestation: JSON.stringify(attestation.value) };
+}
+
+/**
+ * The two seam members that need no boot state, bound once at module scope.
+ *
+ * The chat turn reads its session emit from here, because the creation of a conversation thread
+ * lands inside `prepareChatTurn` and the core bag does not reach that call. Thus one realization
+ * answers both call sites, and a created session carries one claim whichever surface drives it.
+ */
+export const hostProvenanceSeam: ProvenanceSeam = {
+    emitSessionEvent: emitSessionObservation,
+    readExport: readProvenanceExport,
+};
+
+/**
+ * Bind the whole {@link ProvenanceSeam} for the composition root: the run emit of the swappable
+ * holder, beside the two module-scope members above.
+ *
+ * The two model mechanisms stay different on purpose. The run emit carries the name that was stamped
+ * into its inner at construction, and the idle-gated agent switch re-stamps it through
+ * {@link SwappableSandboxEmitters.swap}, so in-flight run work keeps the name it started with. A
+ * session act has no such window: it must name the model at act time, thus the session emit reads
+ * the live source instead.
+ */
+export function createProvenanceSeam(emitters: SwappableSandboxEmitters): ProvenanceSeam {
+    return { ...hostProvenanceSeam, emitRunEvent: emitters.emitProvenance };
 }
