+++ cli/src/modules/harness/runtime.ts
@@ -72,8 +72,7 @@ import { noopExecIngress, startExecIngress, type ExecIngress, type IngressError 
 import { createRunInflexaTool } from "./inflexa_tool.ts";
 import { createLaunchDirTool } from "./launch_dir_tool.ts";
 import { createManageInputsTool } from "./inputs_tool.ts";
-import { createSwappableSandboxEmitters } from "./prov_bridge.ts";
-import { emitReportObservation, installReportSessionModel, readReportProvenance } from "./report_bridge.ts";
+import { createProvenanceSeam, createSwappableSandboxEmitters, installReportSessionModel } from "./prov_bridge.ts";
 import { buildExecuteAnalysisDeps, buildExecuteTargetAssessmentDeps, buildSandboxStepDeps, type RunEngineComposition, type AgentBackend } from "./run_deps.ts";
 import { createUsageRecorder } from "./usage_recorder.ts";
 import { clearAgentSwitch, createSwappableProvider, currentAgentModels, installAgentSwitch } from "./agent_switch.ts";
@@ -1005,6 +1004,10 @@ async function bootHarnessRuntimeOnce(
         // re-points only the cli-owned inner behind them via `emitters.swap`, so
         // no harness-held object is mutated and the swap lands regardless of the workflows' read discipline.
         const emitters = createSwappableSandboxEmitters(`${connection.provider}/${sandboxModel}`);
+        // The whole provenance seam over those emitters: the run emit is the holder's stable handle,
+        // and the session emit and the document read are the module-scope realizations the chat turn
+        // reaches on its own.
+        const provenance = createProvenanceSeam(emitters);
 
         const composition: RunEngineComposition = {
             pool,
@@ -1145,7 +1148,7 @@ async function bootHarnessRuntimeOnce(
                 modelProvider: connection.provider,
                 initialModels: { conversation: conversationModel, sandbox: sandboxModel, utility: utilityModel },
             });
-            // The report bridge names the model that drove each act, and the conversation agent is the
+            // The session emit names the model that drove each act, and the conversation agent is the
             // one that drives a report session. The source reads the switch AT EMIT TIME, thus a live
             // model change re-stamps every later act, the way a swap re-stamps the sandbox emitters.
             // It is installed right after the switch that it reads, thus the two stay in one order. The
@@ -1179,12 +1182,12 @@ async function bootHarnessRuntimeOnce(
                 // what makes the report path exist here at all. The realization runs on the same pinned
                 // runtime that the sandbox uses, thus one boot names one container engine.
                 eyes: createEphemeralEyes({ runtime: pinnedRuntime }),
-                // The two report seams of the signed document. The emit half records each act of a
-                // report session on the bus, and the read half gives the recorded document back for the
-                // page of a preview. The chat turn binds the SAME emit function itself, because the
-                // creation of a conversation thread happens there and this bag does not reach it.
-                emitReportObservation,
-                readReportProvenance,
+                // The one provenance seam of the signed document: the run emit, the session emit that
+                // records each act of a session on the bus, and the read that gives the recorded
+                // document back for the page of a preview. The chat turn binds the SAME module-scope
+                // members itself, because the creation of a conversation thread happens there and this
+                // bag does not reach it.
+                provenance,
             },
             pool: composition.pool,
             skillsDir: cfg.skillsDir,
